* tp/Texinfo/XS/main/build_perl_info.c (store_additional_info): readd cast to intptr_t, as it should be tested on the platforms where it could matter.
* tp/Texinfo/XS/main/extra.c (lookup_info_string): add
lookup_info_string.

* tp/Texinfo/XS/main/builtin_commands.c (element_command_name),
tp/Texinfo/XS/main/convert_to_texinfo.c (expand_cmd_args_to_texi),
tp/Texinfo/XS/parsetexi/close.c (close_brace_command),
tp/Texinfo/XS/parsetexi/parser.c (process_remaining_on_line): use
lookup_info_string.

+++ tp/Texinfo/XS/main/build_perl_info.c
@@ -331,7 +331,7 @@ store_additional_info (const ELEMENT *e, ASSOCIATED_INFO* a, char *key)
               { /* A simple integer.  The intptr_t cast here prevents
                    a warning on MinGW ("cast from pointer to integer of
                    different size"). */
-              IV value = (IV) k->integer;
+              IV value = (IV) (intptr_t) k->integer;
               STORE(newSViv (value));
               break;
               }
@@ -350,7 +350,7 @@ store_additional_info (const ELEMENT *e, ASSOCIATED_INFO* a, char *key)
                   k_integer = lookup_extra (f->contents.list[j], "integer");
                   if (k_integer)
                     {
-                      IV value = (IV) k_integer->integer;
+                      IV value = (IV) (intptr_t) k_integer->integer;
                       av_store (av, j, newSViv (value));
                     }
                   else if (f->contents.list[j]->text.end > 0)
